% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribution.geometric.R
\name{Attribution.geometric}
\alias{Attribution.geometric}
\title{performs category-based geometric attribution}
\usage{
Attribution.geometric(Rp, wp, Rb, wb, wpf = NA, wbf = NA, S = NA,
  Fp = NA, Fb = NA, Rpl = NA, Rbl = NA, Rbh = NA)
}
\arguments{
\item{Rp}{xts of portfolio returns}

\item{wp}{xts of portfolio weights}

\item{Rb}{xts of benchmark returns}

\item{wb}{xts of benchmark weights}

\item{wpf}{vector, xts, data frame or matrix with portfolio weights of 
currency forward contracts}

\item{wbf}{vector, xts, data frame or matrix with benchmark weights of 
currency forward contracts}

\item{S}{(T+1) x n xts, data frame or matrix with spot rates. The first date
should coincide with the first date of portfolio returns}

\item{Fp}{(T+1) x n xts, data frame or matrix with forward rates for contracts in the portfolio. 
The first date should coincide with the first date of portfolio returns}

\item{Fb}{(T+1) x n xts, data frame or matrix with forward rates for contracts in the benchmark. 
The first date should coincide with the first date of benchmark returns}

\item{Rpl}{xts, data frame or matrix of portfolio returns in local currency}

\item{Rbl}{xts, data frame or matrix of benchmark returns in local currency}

\item{Rbh}{xts, data frame or matrix of benchmark returns hedged into the
base currency}
}
\value{
This function returns the list with attribution effects (allocation
or selection effect) including total multi-period attribution effects
}
\description{
Performs category-based geometric attribution of excess return. Calculates 
total geometric attribution effects over multiple periods. Used internally
by the \code{\link{Attribution}} function. 
No interaction effects are computed and the allocation & selection effects are as 
defined by the Brinson Fachler method. In addition, the individual category
effects are NOT adjusted using the possible approach outlined in the Carl Bacon
book. In any case these adjustments are small most of the time. Geometric attribution effects, in 
contrast with arithmetic, do naturally link over time multiplicatively:
\deqn{\frac{(1+R_{p})}{1+R_{b}}-1=\prod^{n}_{t=1}(1+A_{t}^{G})\times
\prod^{n}_{t=1}(1+S{}_{t}^{G})-1}
Total allocation effect at time \eqn{t}:
\deqn{A_{t}^{G}=\frac{1+b_{S}}{1+R_{bt}}-1}
Total selection effect at time \eqn{t}:
\deqn{S_{t}^{G}=\frac{1+R_{pt}}{1+b_{S}}-1}
Semi-notional fund:
\deqn{b_{S}=\sum^{n}_{i=1}w_{pi}\times R_{bi}}
\eqn{w_{pt}}{wpt} - portfolio weights at time \eqn{t},
\eqn{w_{bt}}{wbt} - benchmark weights at time \eqn{t},
\eqn{r_{t}}{rt} - portfolio returns at time \eqn{t},
\eqn{b_{t}}{bt} - benchmark returns at time \eqn{t},
\eqn{r} - total portfolio returns	 
\eqn{b} - total benchmark returns	 
\eqn{n} - number of periods
}
\examples{

data(attrib)
Attribution.geometric(Rp = attrib.returns[, 1:10], wp = attrib.weights[1, ],
Rb = attrib.returns[, 11:20], wb = attrib.weights[2, ])

}
\references{
Christopherson, Jon A., Carino, David R., Ferson, Wayne E.  
\emph{Portfolio Performance Measurement and Benchmarking}. McGraw-Hill. 
2009. Chapter 18-19 \cr Bacon, C. \emph{Practical Portfolio Performance 
Measurement and Attribution}. Wiley. 2004. Chapter 5, 8, Appendix A \cr
}
\seealso{
\code{\link{Attribution}}
}
\author{
Andrii Babii
}
\keyword{attribution,}
\keyword{geometric}
\keyword{linking}
