% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribution.levels.R
\name{Attribution.levels}
\alias{Attribution.levels}
\title{provides multi-level sector-based geometric attribution}
\usage{
Attribution.levels(Rp, wp, Rb, wb, Rpl = NA, Rbl = NA, Rbh = NA, h,
  h_levels, geometric = FALSE, anchored = TRUE)
}
\arguments{
\item{Rp}{xts, data frame or matrix of portfolio returns}

\item{wp}{vector, xts, data frame or matrix of portfolio weights}

\item{Rb}{xts, data frame or matrix of benchmark returns}

\item{wb}{vector, xts, data frame or matrix of benchmark weights}

\item{Rpl}{xts, data frame or matrix of portfolio returns in local currency}

\item{Rbl}{xts, data frame or matrix of benchmark returns in local currency}

\item{Rbh}{xts, data frame or matrix of benchmark returns hedged into the
base currency}

\item{h}{data.frame with the hierarchy obtained from the buildHierarchy 
function or defined manually in the same style as buildHierarchy's
output}

\item{h_levels}{The remaining passthrough parameters represent the levels in the
hierarchy to aggregate by}

\item{geometric}{TRUE/FALSE,  whether to use geometric or arithmetic excess
returns for the attribution analysis. By default this is set to FALSE, which results
in arithmetic excess return attribution.}

\item{anchored}{TRUE/FALSE, to indicate if the weights at each level should be
anchored based on prior level's decision, as outlined in the Morningstar
methodology documented referenced below}
}
\value{
returns the list with geometric excess returns including annualized
geometric excess returns, total attribution effects (allocation, selection 
and total) including total multi-period attribution effects, attribution 
effects at each level and security selection
}
\description{
Provides multi-level sector-based geometric attribution. The Brinson model
attributes excess returns at one level. This function works with more 
complex decision processes. For instance, the 3-level decision process 
may have the following levels: type of asset - country - sector. The 
levels should be specified in the vector with elements in the particular 
order: from the highest level to the lowest. Returns and weighs for 
portfolio and benchmark should be at the lowest level (e.g. individual 
instruments). Benchmark should have the same number of columns as portfolio.
That is there should be a benchmark for each instrument in the portfolio 
(possibly 0). The contribution to the allocation in the \eqn{i^{th}}
category for the \eqn{d^{th}} level is: 
\deqn{\left(^{d}w_{pi}-^{d}w_{bi}\right)\times
\left(\frac{1+^{d}R_{bi}}{1+^{d-1}R_{bi}}-1\right)
\times\frac{1+^{d-1}R_{bi}}{1+bs^{d-1}}}
The total attribution for each asset allocation step in the decision process
is: \deqn{\frac{1+^{d}bs}{1+^{d-1}bs}-1}
The final step, stock selection, is measured by:
\deqn{^{d}w_{pi}\times\left(\frac{1+R_{pi}}{1+^{d}R_{bi}}-1\right)
\times\frac{1+^{d}R_{bi}}{1+^{d}bs}}
}
\examples{

data(attrib)
Attribution.levels(Rp = attrib.returns[, 1:10], wp = attrib.weights[1, ], Rb = attrib.returns[, 11:20], 
wb = attrib.weights[2, ], h = attrib.hierarchy, c("type", "MarketCap", "Sector"))
Attribution.levels(Rp = attrib.returns[, 1:10], wp = attrib.weights[1, ], Rb = attrib.returns[, 11:20], 
wb = attrib.weights[2, ], h = attrib.hierarchy, c("type", "Sector"))

}
\references{
Bacon, C. \emph{Practical Portfolio Performance Measurement and
Attribution}. Wiley. 2004. p. 215-220

\url{https://corporate.morningstar.com/us/documents/MethodologyDocuments/MethodologyPapers/EquityPerformanceAttributionMeth.pdf}
}
\seealso{
\code{\link{Attribution.geometric}}
}
\author{
Andrii Babii
}
\keyword{attribution}
\keyword{attribution,}
\keyword{geometric}
\keyword{multi-level}
