% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribution.R
\name{Attribution}
\alias{Attribution}
\title{performs single-category attribution}
\usage{
Attribution(
  Rp,
  wp,
  Rb,
  wb,
  wpf = NA,
  wbf = NA,
  S = NA,
  Fp = NA,
  Fb = NA,
  Rpl = NA,
  Rbl = NA,
  Rbh = NA,
  bf = TRUE,
  method = "none",
  currency_method = "none",
  linking = "grap",
  geometric = FALSE,
  contribution = FALSE,
  adjusted = FALSE,
  annualization = "none",
  annualization_scale = NA,
  impute_returns = TRUE
)
}
\arguments{
\item{Rp}{T x n xts, data frame or matrix of portfolio returns}

\item{wp}{vector, xts, data frame or matrix of portfolio weights}

\item{Rb}{T x n xts, data frame or matrix of benchmark returns}

\item{wb}{vector, xts, data frame or matrix of benchmark weights}

\item{wpf}{vector, xts, data frame or matrix with portfolio weights of 
currency forward contracts}

\item{wbf}{vector, xts, data frame or matrix with benchmark weights of 
currency forward contracts}

\item{S}{(T+1) x n xts, data frame or matrix with spot rates. The first date
should coincide with the first date of portfolio returns}

\item{Fp}{(T+1) x n xts, data frame or matrix with forward rates for contracts in the portfolio. 
The first date should coincide with the first date of portfolio returns}

\item{Fb}{(T+1) x n xts, data frame or matrix with forward rates for contracts in the benchmark. 
The first date should coincide with the first date of benchmark returns}

\item{Rpl}{xts, data frame or matrix of portfolio returns in local currency}

\item{Rbl}{xts, data frame or matrix of benchmark returns in local currency}

\item{Rbh}{xts, data frame or matrix of benchmark returns hedged into the (only used when geometric is set to TRUE)
base currency}

\item{bf}{TRUE for Brinson and Fachler and FALSE for Brinson, Hood and 
Beebower arithmetic attribution. By default this set to TRUE and so Brinson and Fachler 
attribution is selected. This is only used for Arithmetic attribution while Geometric
attribution only supports Brinson Fachler.}

\item{method}{Used to select the priority between allocation and selection 
effects in arithmetic attribution. May be any of: \itemize{ \item none - 
present allocation, selection and interaction effects independently, 
\item top.down - the priority is given to the group allocation. Interaction
term is combined with the security selection effect, \item bottom.up - the 
priority is given to the security selection. Interaction term is combined 
with the group allocation effect} 
By default "none" is selected}

\item{currency_method}{Used to select the approach to use to compute currency attribution effects
May be any of: \itemize{ \item none - 
no currency attribution effects are computed, 
\item standard - currency attribution effects are computed as the difference between the attribution effects
computed with base and local currencies
\item ankrim.hensel - currency attribution effects are computed using Ankrim-Hensel methodology (single-period only)}
By default "none" is selected. Note that, at the moment, multi-currency attribution is not supported for Davies & Laker multi-period linking
and the Ankrim-Hensel method is only implemented for single-period attribution.}

\item{linking}{Used to select the linking method to present the multi-period
summary of arithmetic attribution effects. May be any of: 
\itemize{\item carino - logarithmic linking coefficient method
\item menchero - Menchero's smoothing algorithm
\item grap - linking approach developed by GRAP
\item frongello - Frongello's linking method
\item davies.laker - Davies and Laker's linking method}
By default grap linking is selected. This option is ignored if 'geometric' is set to TRUE or
if the data does not imply multi-period attribution. Davies & Laker does not support multi-currency attribution
at the moment.}

\item{geometric}{TRUE/FALSE, whether to use geometric or arithmetic excess
returns for the attribution analysis. By default this is set to FALSE, which results
in arithmetic excess return attribution.}

\item{contribution}{TRUE/FALSE, whether to also compute and return portfolio & benchmark contributions
for each category and period. Defaults to FALSE.}

\item{adjusted}{TRUE/FALSE, whether to show original or smoothed attribution
effects for each period. By default unadjusted attribution effects are 
returned (this is not used for Davies and Laker's linking method for arithmetic attribution 
or for geometric attribution as it is not applicable in either case)}

\item{annualization}{Used to select the annualization method for multi-period returns, excess returns, contributions
and attribution effects. May be any of:
\itemize{\item none - no annualized numbers are returned
\item standard 
\item proportional}
The default is 'none'. See vignette on annualization for detailed descriptions of the supported methods}

\item{annualization_scale}{Used to define a custom scale factor to use when annualizing returns. This is only
applicable if the periodicity of the data is daily and otherwise this parameter is ignored. May be any of:
\itemize{\item NA - use built-in scale factor
\item <numeric> - any numeric value
\item "monthly" - count the periods in terms of months (including fractional months) to determine scale}
The default is 'NA'. See vignette on annualization for more details.}

\item{impute_returns}{TRUE/FALSE, whether to impute returns from the benchmark, when the weight of the asset
is zero in the portfolio & vice-versa. These imputed returns are only used in the calculations
for attributions effects but not for cumulative asset returns (annualized or otherwise). Defaults to TRUE.}
}
\value{
returns a list with the following components: excess returns with
annualized excess returns over all periods, attribution effects (allocation, 
selection and interaction) and optionally, portfolio and benchmark contributions 
for each category and period. Each element of the list is structured as a data.frame for consistency purposes.
}
\description{
Performs single-category attribution analysis. Portfolio 
performance measured relative to a benchmark gives an indication of the 
value-added by the portfolio. Equipped with weights and returns of portfolio
segments, we can dissect the value-added into useful components. This 
function is based on the group-based approach to the attribution. The 
workhorse is the Brinson model that explains the arithmetic difference 
between portfolio and benchmark returns. That is it breaks down the 
arithmetic excess returns at one category level. If returns and weights are available
at the lowest category level (e.g. for individual instruments), the aggregation up to
the chosen category from the hierarchy can be done using 
\code{\link{Return.level}} function. The attribution effects can be computed
for several periods. The multi-period summary is obtained using one of 
linking methods: Carino, Menchero, GRAP, Frongello or Davies Laker. It also
allows to break down the geometric excess returns, which link naturally over
time. Finally, it supports annualization of excess returns, attribution effects and contributions.
}
\details{
The arithmetic excess returns are decomposed into the sum of allocation, 
selection and interaction effects across \eqn{n} groups:
\deqn{R_{p}-R_{b}=\sum^{n}_{i=1}\left(A_{i}+S_{i}+I_{i}\right)}{R_p - R_b = \sum {i=1 to n} (A_i + S_i + I_i)}
The arithmetic attribution effects for the category i are computed
as suggested in the Brinson, Hood and Beebower (1986):
Allocation effect
\deqn{A_{i}=(w_{pi}-w_{bi})\times R_{bi}}{A_i = (w_pi - w_bi) * R_bi}
Selection effect
\deqn{S_{i}=w_{bi}\times(R_{pi}-R_{bi})}{S_i = w_bi * (R_pi - R_bi)}
Interaction effect
\deqn{I_{i}=(w_{pi}-w_{bi})
\times(R_{pi}-R_{bi})}{I_i = (w_pi - w_bi) * (R_pi - R_bi)}
\eqn{R_{p}}{R_p} - total portfolio returns,
\eqn{R_{b}}{R_b} - total benchmark returns, 
\eqn{w_{pi}}{w_pi} - weights of the category \eqn{i} in the portfolio, 
\eqn{w_{bi}}{w_bi} - weights of the category \eqn{i} in the benchmark, 
\eqn{R_{pi}}{R_pi} - returns of the portfolio category \eqn{i}, 
\eqn{R_{bi}}{R_bi} - returns of the  benchmark category \eqn{i}.
If Brinson and Fachler (1985) is selected the allocation effect differs:
\deqn{A_{i}=(w_{pi}-w_{bi})
\times (R_{bi} - R_{b})}{A_i = (w_pi - w_bi) * (R_bi - R_b)}
Depending on goals we can give priority to the allocation or to 
the selection effects. If the priority is given to the group allocation
the interaction term will be combined with the security selection effect
(top-down approach). If the priority is given to the security selection,
the interaction term will be combined with the asset-allocation effect
(bottom-up approach).
Usually we have more than one period. In that case individual arithmetic 
attribution effects should be adjusted using linking methods. Adjusted
arithmetic attribution effects can be summed up over time to provide the
multi-period summary: 
\deqn{R_{p}-R_{b}=\sum^{T}_{t=1}\left(A_{t}'+S_{t}'+I_{t}'\right)}{R_p - R_b = \sum {t=1 to T} (A_t'+S_t'+I_t')}
where \eqn{T} is the number of periods and prime stands for the adjustment 
(this is not applicable to Davies & Laker linking method).
The geometric attribution effects do not suffer from the linking problem.
Moreover we don't have the interaction term. For more details about the 
geometric attribution see the documentation to 
\code{\link{Attribution.geometric}}. Finally, arithmetic annualized excess 
returns are computed as the arithmetic difference between annualised 
portfolio and benchmark returns:
\deqn{E^A_ann = R_{pa}-R_{ba}}{E^A_ann = R_pa - R_ba} and the geometric annualized excess
returns are computed as the geometric difference between annualized 
portfolio and benchmark returns: 
\deqn{E^G_ann =\frac{1+R_{pa}}{1+R_{ba}}-1}{E^G_ann = (1 + R_pa) / (1 + R_ba) - 1}
In the case of \strong{arithmetic attribution} of a multi-currency portfolio, the currency return, currency
surprise and forward premium should be specified. The multi-currency
arithmetic attribution is handled following either a standard or Ankrim and Hensel approach (1992).
With the standard approach, the currency effects are simply the difference between the attribution effects computed
with the base currency and the local currency.
With Ankrim and Hensel, the currency returns are decomposed into the sum of the currency surprise and
the forward premium: \deqn{R_{ci} = R_{ei} + R_{di}}{R_ci = R_ei + R_di}
where 
\deqn{R_{ei} = \frac{S_{i}^{t+1} - F_{i}^{t+1}}{S_{i}^{t}}}{R_ei = (S_i^(t+1) - F_i^(t+1)) / S_i^t}
\deqn{R_{di} = \frac{F_{i}^{t+1}}{S_{i}^{t}} - 1}{R_di = (F_i^(t+1) / S_i^t) - 1}
\eqn{S_i^t}{S_i^t} - spot rate for asset \eqn{i} at time \eqn{t},
\eqn{S_i^{t+1}}{S_i^(t+1)} - spot rate for asset \eqn{i} at time \eqn{t+1} and
\eqn{F_i^{t+1}}{F_i^(t+1)} - forward rate for asset \eqn{i} at time \eqn{t} for converstion at time \eqn{t+1}. 
Excess returns are decomposed into the sum of allocation, selection and 
interaction effects as in the standard Brinson model: 
\deqn{R_{p}-R_{b}=\sum^{n}_{i=1}\left(A_{i}+S_{i}+I_{i}\right)}{R_p - R_ b = \sum {i=1 to n} (A_i + S_ i + I_i)}
However the allocation effect is computed taking into account currency
effects:
\deqn{A_{i}=(w_{pi}-w_{bi})\times (R_{bi} - R_{ci} - R_{l})}{A_i = (w_pi - w_bi) * (R_bi - R_ci - R_l)}
Benchmark returns adjusted to the currency:
\deqn{R_{l} = \sum^{n}_{i=1}w_{bi}\times(R_{bi}-R_{ci})}{R_l = \sum {i=1 to n} (w_bi * (R_bi - R_ci))}
The contribution from the currency is analogous to asset allocation:
\deqn{C_{i} = (w_{pi} - w_{bi}) \times (R_{ei} - R_e) + (w_{pfi} - w_{bfi}) \times (R_{fi} - R_e)}{C_i = (w_pi - w_bi) * (R_ei - R_e) + (w_pfi - w_bfi) * (R_fi - R_e)}
where \deqn{R_e = \sum^{n}_{i=1}w_{bi}\times R_{ei}}{R_e = \sum {i=1 to n} (w_bi * R_ei)}
The final term, forward premium, is also analogous to the asset allocation:
\deqn{D_{i} = (w_{pi} - w_{bi}) \times (R_{di} - R_d)}{D_i = (w_pi - w_bi) * (R_di - R_d)}
where \deqn{R_d = \sum^{n}_{i=1}w_{bi}\times R_{di}}{R_d = \sum {i=1 to n} (w_bi * R_di)}
and \eqn{R_{di}}{R_di} is the forward premium.
In general if the intent is to estimate statistical parameters, the 
arithmetic excess return is preferred. However, due to the linking 
challenges, it may be preferable to use geometric excess return if the 
intent is to link and annualize excess returns.
}
\examples{

data(sample_data)
Attribution(Rp = multi_period_portf_2$Rp, wp = multi_period_portf_2$wp, 
            Rb = multi_period_portf_2$Rb, wb = multi_period_portf_2$wb, 
            method = "top.down", linking = "grap")

}
\references{
Ankrim, E. and Hensel, C. \emph{Multi-currency performance
attribution}. Russell Research Commentary. November 2002 \cr Bacon, C. 
\emph{Practical Portfolio Performance Measurement and Attribution}. Wiley.
2004. Chapter 5, 6, 8 \cr Christopherson, Jon A., Carino, David R., Ferson, 
Wayne E. \emph{Portfolio Performance Measurement and Benchmarking}. 
McGraw-Hill. 2009. Chapter 18-19 \cr Brinson, G. and Fachler, N. (1985)
\emph{Measuring non-US equity portfolio performance}. Journal of Portfolio
Management. Spring. p. 73 -76. \cr Gary P. Brinson, L. Randolph Hood, and 
Gilbert L. Beebower, \emph{Determinants of Portfolio Performance}. Financial
Analysts Journal. vol. 42, no. 4, July/August 1986, p. 39-44 \cr 
Karnosky, D. and Singer, B. \emph{Global asset management and performance 
attribution. The Research Foundation of the Institute of Chartered Financial
Analysts}. February 1994. \cr
}
\seealso{
\code{\link{Attribution.levels}}, 
\code{\link{Attribution.geometric}}
}
\author{
Andrii Babii
}
\keyword{attribution}
