% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMonthlyPeriods}
\alias{getMonthlyPeriods}
\title{Compute periods in increment of months}
\usage{
getMonthlyPeriods(start_date, end_date)
}
\arguments{
\item{start_date}{Date object representing the starting date}

\item{end_date}{Date object representing the ending date}
}
\value{
numeric representing the fractional number of months between the two dates provided
}
\description{
Computes the total number of months between two dates including fractional months
}
\author{
Anand Krishnakumar
}
