% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{multi_period_portf_4}
\alias{multi_period_portf_4}
\title{Fourth multi-period, multi-currency portfolio & benchmark example with country-level data, for use in vignettes & unit tests}
\format{
A list with 6 members:
\describe{
  \item{Rp}{xts object with portfolio base currency returns for each country and period}
  \item{Rb}{xts object with benchmark base currency returns for each country and period}
  \item{wp}{xts object with portfolio weights for each country and period}
  \item{wb}{xts object with benchmark weights for each country and period}
  \item{Rpl}{xts object with portfolio local currency returns for each country and period}
  \item{Rbl}{xts object with benchmark local currency returns for each country and period}
}
}
\usage{
multi_period_portf_4
}
\description{
A test data set containing portfolio & benchmark returns for a few countries over a number of quarters 
For each period we have portfolio & benchmark weights, base and local returns for each country.
}
\keyword{datasets}
