% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{multi_period_portf_5}
\alias{multi_period_portf_5}
\title{Fifth multi-period portfolio & benchmark example with security-level data, for use in vignettes & unit tests}
\format{
A list with 4 members:
\describe{
  \item{Rp}{xts object with portfolio returns for each security and period}
  \item{Rb}{xts object with benchmark returns for each security and period}
  \item{wp}{xts object with portfolio weights for each security and period}
  \item{wb}{xts object with benchmark weights for each security and period}
}
}
\usage{
multi_period_portf_5
}
\description{
A test data set containing portfolio & benchmark returns for a large number of securities over several periods. 
For each period we have portfolio & benchmark weights and returns for each security.
}
\keyword{datasets}
