% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{simple_multi_curr_portf_1}
\alias{simple_multi_curr_portf_1}
\title{Simple multi-currency portfolio & benchmark example, categorized into sectors, for use in vignettes & unit tests}
\format{
An object of class \code{data.frame} with 6 rows and 11 columns.
}
\usage{
simple_multi_curr_portf_1
}
\description{
A simple portfolio & benchmark test data set, categorized into sectors, that contains portfolio & benchmark weights and 
returns for each category in both local and base currency terms.
}
\keyword{datasets}
