% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{simple_multi_curr_portf_2}
\alias{simple_multi_curr_portf_2}
\title{Simple multi-currency portfolio & benchmark example, categorized into regions, for use in vignettes & unit tests}
\format{
An object of class \code{data.frame} with 6 rows and 3 columns.
}
\usage{
simple_multi_curr_portf_2
}
\description{
A simple portfolio & benchmark test data set, categorized into regions, that contains portfolio & benchmark weights and 
returns for each category, portfolio & benchmark forward contract weights and spot ares and forward rates for the currencies 
corresponding to each region.
}
\keyword{datasets}
