% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{simple_portf_1}
\alias{simple_portf_1}
\title{Simple three category portfolio & benchmark example for use in vignettes}
\format{An object of class \code{data.frame} with 4 rows and 3 columns.}
\usage{
simple_portf_1
}
\description{
A simple three category portfolio & benchmark test data set that contains portfolio & benchmark weights and returns.
}
\keyword{datasets}
