% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{simple_portf_2}
\alias{simple_portf_2}
\title{Simple portfolio & benchmark example, categorized into sectors, for use in vignettes & unit tests}
\format{
An object of class \code{data.frame} with 4 rows and 10 columns.
}
\usage{
simple_portf_2
}
\description{
A simple portfolio & benchmark test data set, categorized into sectors, that contains portfolio & benchmark weights and returns for each category.
}
\keyword{datasets}
