% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribution.geometric.R
\name{Attribution.geometric}
\alias{Attribution.geometric}
\title{performs category-based geometric attribution}
\usage{
Attribution.geometric(
  Rp,
  wp,
  Rb,
  wb,
  wpf = NA,
  wbf = NA,
  S = NA,
  Fp = NA,
  Fb = NA,
  Rpl = NA,
  Rbl = NA,
  Rbh = NA,
  contribution = FALSE,
  annualization = "none",
  annualization_scale = NA,
  impute_returns = TRUE
)
}
\arguments{
\item{Rp}{xts of portfolio returns}

\item{wp}{xts of portfolio weights}

\item{Rb}{xts of benchmark returns}

\item{wb}{xts of benchmark weights}

\item{wpf}{vector, xts, data frame or matrix with portfolio weights of 
currency forward contracts}

\item{wbf}{vector, xts, data frame or matrix with benchmark weights of 
currency forward contracts}

\item{S}{(T+1) x n xts, data frame or matrix with spot rates. The first date
should coincide with the first date of portfolio returns}

\item{Fp}{(T+1) x n xts, data frame or matrix with forward rates for contracts in the portfolio. 
The first date should coincide with the first date of portfolio returns}

\item{Fb}{(T+1) x n xts, data frame or matrix with forward rates for contracts in the benchmark. 
The first date should coincide with the first date of benchmark returns}

\item{Rpl}{xts, data frame or matrix of portfolio returns in local currency}

\item{Rbl}{xts, data frame or matrix of benchmark returns in local currency}

\item{Rbh}{xts, data frame or matrix of benchmark returns hedged into the
base currency}

\item{contribution}{TRUE/FALSE, whether to also compute and return portfolio & benchmark contributions
for each category and period. Defaults to FALSE.}

\item{annualization}{Used to select the annualization method for multi-period excess returns, contributions
and attribution effects. May be any of:
\itemize{\item none - no annualized numbers are returned
\item standard 
\item proportional}
The default is 'none'. See vignette on annualization for detailed descriptions of the supported methods}

\item{annualization_scale}{Used to define a custom scale factor to use when annualizing returns. This is only
applicable if the periodicity of the data is daily and otherwise this parameter is ignored. May be any of:
\itemize{\item NA - use built-in scale factor
\item <numeric> - any numeric value
\item "monthly" - count the periods in terms of months (including fractional months) to determine scale}}

\item{impute_returns}{TRUE/FALSE, whether to impute returns from the benchmark, when the weight of the asset
is zero in the portfolio & vice-versa. These imputed returns are only used in the calculations
for attributions effects but not for cumulative asset returns (annualized or otherwise). Defaults to TRUE.}
}
\value{
This function returns the list with attribution effects (allocation
or selection effect) including total multi-period attribution effects
}
\description{
Performs category-based geometric attribution of excess return. Calculates 
total geometric attribution effects over multiple periods. Used internally
by the \code{\link{Attribution}} function. 
No interaction effects are computed and the allocation & selection effects are as 
defined by the Brinson Fachler method. In addition, the individual category
effects are NOT adjusted using the possible approach outlined in the Carl Bacon
book. In any case these adjustments are small most of the time. Geometric attribution effects, in 
contrast with arithmetic, do naturally link over time multiplicatively.
}
\details{
The total geometric excess return over time is linked to the periodic geometric effects as follows:
\deqn{\frac{(1+R_{p})}{1+R_{b}}-1=\prod^{n}_{t=1}(1+A_{t}^{G}) \times \prod^{n}_{t=1}(1+S_{t}^{G})-1}{(1 + R_p) / (1 + R_b) - 1 = \prod {t=1 to n} (1 + A_t^G) * \prod {t=1 to n} (1+S_t^G) - 1}
where the total allocation effect at time \eqn{t} is:
\deqn{A_{t}^{G}=\frac{1+R_{bst}}{1+R_{bt}}-1}{A_t^G = (1 + R_bst) / (1 + R_bt) - 1}
and the total selection effect at time \eqn{t} is:
\deqn{S_{t}^{G}=\frac{1+R_{pt}}{1+R_{bst}}-1}{S_t^G = (1 + R_pt) / (1 + R_bst) - 1}
where the semi-notional fund returns at time \eqn{t} is:
\deqn{R_{bst}=\sum^{n}_{i=1}w_{pit}\times R_{bit}}{R_bst = \sum {i=1 to n} (w_pit * R_bit)}
with
\eqn{w_{pt}}{w_pt} being the portfolio weights at time \eqn{t},
\eqn{w_{bt}}{w_bt} being the benchmark weights at time \eqn{t},
\eqn{R_{pt}}{R_pt} being the portfolio returns at time \eqn{t},
\eqn{R_{bt}}{R_bt} being benchmark returns at time \eqn{t},
\eqn{R_{p}}{R_p} being the total portfolio returns,
\eqn{R_{b}}{R_b} being the total benchmark returns	and 
\eqn{n} being the number of periods.

The multi-currency geometric attribution is handled following the Appendix B
(Bacon, 2004). 

The individual selection effects are computed using:
\deqn{S_i = w_{pi}\times\left(\frac{1+R_{pli}}{1+R_{bli}}-1\right)\times \left(\frac{1+R_{bli}}{1+R_{bsl}}\right)}{S_i = w_pi * ((1 + R_pli) / (1 + R_bli) - 1) * ((1 + R_bli) / (1 + R_bsl))}

The individual allocation effects are computed using:
\deqn{A_i = (w_{pi}-w_{bi})\times\left(\frac{1+R_{bhi}}{1+R_{bl}}-1\right)}{A_i = (w_pi - w_bi) * ((1 + R_bhi) / (1 + R_bl) - 1)}

where the total semi-notional returns hedged into the base currency is:
\deqn{R_{bsh} = \sum^{n}_{i=1}(((w_{pi} - w_{bi})\times R_{bhi}) + w_{bi}\times R_{bli})}{R_bsh = \sum {i=1 to n} (((w_pi - w_bi) * R_bhi) + w_bi * R_bli)}
and the total semi-notional returns in the local currency is given by:
\deqn{R_{bsl} = \sum^{n}_{i=1}(w_{pi} \times R_{bli})}{R_bsl = \sum {i=1 to n} (w_pi * R_bli)}
with \eqn{R_{pli}}{R_pli} being the portfolio returns in the local currency,
\eqn{R_{bli}}{R_bli} being the benchmark returns in the local currency,
\eqn{R_{bhi}}{R_bhi} being the benchmark returns hedged into the base currency,
\eqn{R_{bl}}{R_bl} being the total benchmark returns in the local currency and
\eqn{R_{pl}}{R_pl} being the total portfolio returns in the local currency

The total excess returns are decomposed into:
\deqn{\frac{(1+R_{p})}{1+R_{b}}-1=\frac{1+R_{bsh}}{1+R_{bl}}\times\frac{1+R_{pl}}{1+R_{bsl}}\times\frac{1+R_{p}}{1+R_{pl}}\times\frac{1+R_{bl}}{1+R_{b}}\times\frac{1+R_{bsl}}{1+R_{bsh}}-1}{(1 + R_p)  / (1 + R_b) - 1 = ((1 + R_bsh) / (1 + R_bl)) * ((1 + R_pl) / (1 + R_bsl)) * ((1 + R_p) / (1 + R_pl)) * ((1 + R_bl) / (1 + R_b)) * ((1 + R_bsl) / (1 + R_bsh)) - 1}

where the first term corresponds to the allocation, second to the selection,
the next two to the naive currency attribution effect and the last to the hedging cost transferred.
}
\examples{

data(sample_data)
Attribution.geometric(Rp = multi_period_portf_2$Rp, wp = multi_period_portf_2$wp, 
                      Rb = multi_period_portf_2$Rb, wb = multi_period_portf_2$wb)

}
\references{
Christopherson, Jon A., Carino, David R., Ferson, Wayne E.  
\emph{Portfolio Performance Measurement and Benchmarking}. McGraw-Hill. 
2009. Chapter 18-19 \cr Bacon, C. \emph{Practical Portfolio Performance 
Measurement and Attribution}. Wiley. 2004. Chapter 5, 8, Appendix A \cr
}
\seealso{
\code{\link{Attribution}}
}
\author{
Andrii Babii
}
\keyword{attribution,}
\keyword{geometric}
\keyword{linking}
